the work done in lab today

+++ final/Control/feedback.h
@@ -1,10 +1,10 @@
 #ifndef FEEDBACK_H
 #define FEEDBACK_H
 
-void initialize_feedback(unsigned long time);
+void initialize_feedback(double time);
 
-float do_feedback_forward(double distance, unsigned long time);
-float do_feedback_turn(double phi, unsigned long time);
+float do_feedback_forward(double distance, double time);
+float do_feedback_turn(double phi, double time);
 
 #endif
